Fix endpoints

+++ internal/storage/postgresql/storage.go
@@ -30,8 +30,8 @@ func NewStorage(pgConfig config.PostgresServer, logger Logger.Logger) (*Storage,
 	return &Storage{conn, logger}, nil
 }
 
-func (s *Storage) SelectAll() ([]string, error) {
-	rows, err := s.db.Query(`SELECT * FROM urls`)
+func (s *Storage) SelectAllUserId(userId uint64) ([]string, error) {
+	rows, err := s.db.Query("SELECT * FROM urls WHERE user_id = $1", userId)
 	if err != nil {
 		return nil, err
 	}
@@ -42,14 +42,15 @@ func (s *Storage) SelectAll() ([]string, error) {
 		var id int64
 		var url string
 		var alias string
-		err = rows.Scan(&id, &alias, &url)
+		var userId uint64
+		err = rows.Scan(&id, &alias, &url, &userId)
 		if err != nil {
 			return nil, err
 		}
-		res = append(res, strconv.Itoa(int(id))+" "+url+" "+alias)
+		res = append(res, strconv.Itoa(int(id))+", "+url+", "+alias+", "+strconv.Itoa(int(userId)))
 	}
 
-	s.logger.Debug("Returning from SelectAll")
+	s.logger.Debug("Returning from SelectAllUserId")
 
 	return res, nil
 }
@@ -74,16 +75,16 @@ func (s *Storage) GetUniqueFreeAlias(aliasLen int) (string, error) {
 	return randString, nil
 }
 
-func (s *Storage) AddUrl(url string, alias string) (*entities.URL, error) {
+func (s *Storage) AddUrl(url string, alias string, userId uint64) (*entities.URL, error) {
 	var id uint64
-	err := s.db.Get(&id, "INSERT INTO urls (url, alias) VALUES($1, $2) RETURNING id", url, alias)
+	err := s.db.Get(&id, "INSERT INTO urls (url, alias, user_id) VALUES($1, $2, $3) RETURNING id", url, alias, userId)
 
 	if err != nil {
 		s.logger.Error("Error with insert", err)
 		return nil, err
 	}
 
-	return &entities.URL{Id: id, Url: url, Alias: alias}, nil
+	return &entities.URL{Id: id, Url: url, Alias: alias, UserId: userId}, nil
 }
 
 func (s *Storage) GetUrlByAlias(alias string) (*entities.URL, error) {
@@ -101,8 +102,9 @@ func (s *Storage) GetUrlByAlias(alias string) (*entities.URL, error) {
 	return &urls[0], nil
 }
 
-func (s *Storage) DeleteUrlByAlias(alias string) (bool, error) {
-	result, err := s.db.Exec("DELETE FROM urls WHERE alias=$1 RETURNING url", alias)
+func (s *Storage) DeleteUrlByAlias(alias string, userId uint64) (bool, error) {
+
+	result, err := s.db.Exec("DELETE FROM urls WHERE alias=$1 AND user_id=$2 RETURNING url", alias, userId)
 	if err != nil {
 		s.logger.Error("Error with delete", err)
 		return false, err
